requirements.txt
cartopy
matplotlib
netcdf4
numpy
pandas
python-dotenv
requests
scipy
validators
xarray
pyyaml